Remove folder in controller
