+++ routes/web.php
@@ -1,8 +1,8 @@
 <?php
 
+use App\Http\Controllers\Cms\DeptController;
+use App\Http\Controllers\Cms\NewsController;
 use App\Http\Controllers\Contoh\ContohController;
-use App\Http\Controllers\DepartementController;
-use App\Http\Controllers\News\NewsController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -26,9 +26,9 @@ Route::prefix('News')->group(function () {
 });
 
 Route::prefix('Departement')->group(function () {
-    Route::get('index', [DepartementController::class, 'index'])->name('departement.index');
-    Route::post('create', [DepartementController::class, 'create'])->name('departement.create');
-    Route::get('getspecdata/{id}', [DepartementController::class, 'edit']);
-    Route::post('update', [DepartementController::class, 'update'])->name('departement.update');
-    Route::delete('deletespecdata/{id}', [DepartementController::class, 'delete']);
+    Route::get('index', [DeptController::class, 'index'])->name('departement.index');
+    Route::post('create', [DeptController::class, 'create'])->name('departement.create');
+    Route::get('getspecdata/{id}', [DeptController::class, 'edit']);
+    Route::post('update', [DeptController::class, 'update'])->name('departement.update');
+    Route::delete('deletespecdata/{id}', [DeptController::class, 'delete']);
 });
